Fix backend trying to remove a non existing picture

+++ gifts_backend/models.py
@@ -139,7 +139,8 @@ class Gift(db.Model):
 
 @event.listens_for(Gift, 'after_delete')
 def delete_picture(mapper, connection, target):
-    os.remove(target.picture)
+    if target.picture is not None and target.picture != '':
+        os.remove(target.picture)
 
 
 class Comment(db.Model):
